Add struct components

+++ product/src/lib.rs
@@ -3,7 +3,7 @@ use serde::{Deserialize, Serialize};
 #[derive(Deserialize, Serialize, Debug, PartialEq, Clone)]
 pub struct Product {
     pub name: String,
-    pub price: f32,
+    pub price: i32,
     pub quantity: i32,
 }
